feat: Update Home component to fix drag value assignment
The Home component has been updated to fix the assignment of the drag value in the onScroll event handler. Previously, the drag value was set to 0 when the contentOffset.y was not equal to 0, which was incorrect. This commit corrects the logic by setting the drag value to 1 when the contentOffset.y is not equal to 0. This change ensures that the drag value is properly updated based on the scroll position, improving the functionality of the Home component.